fix(vike): rollback vike to vite to avoid breaking change and reuse code on @tsed/vite-plugin-ssr
