+++ packages/third-parties/vike/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@tsed/vike",
-  "version": "7.60.0",
-  "description": "Vite SSR plugin package for Ts.ED framework",
+  "version": "7.60.1",
+  "description": "Vike package for Ts.ED framework",
   "private": false,
   "contributors": [
     {
@@ -30,24 +30,27 @@
     "sirv": ">=2.0.2"
   },
   "devDependencies": {
-    "@tsed/common": "7.60.0",
-    "@tsed/core": "7.60.0",
-    "@tsed/di": "7.60.0",
+    "@tsed/common": "7.60.1",
+    "@tsed/core": "7.60.1",
+    "@tsed/di": "7.60.1",
     "@tsed/eslint": "7.44.1",
-    "@tsed/exceptions": "7.60.0",
+    "@tsed/exceptions": "7.60.1",
     "@tsed/typescript": "7.44.1",
     "eslint": "^8.12.0",
     "vite": "5.0.12",
     "vike": "0.4.160"
   },
   "peerDependencies": {
-    "@tsed/common": "^7.60.0",
+    "@tsed/common": "^7.60.1",
     "vite": ">=4",
     "vike": ">=0.4.160"
   },
   "peerDependenciesMeta": {
     "@tsed/common": {
       "optional": false
+    },
+    "vike": {
+      "optional": true
     }
   }
 }
