+++ packages/third-parties/vike/src/ViteModule.ts
