--- packages/third-parties/vike/src/decorators/vike.ts
@@ -1,5 +1,0 @@
-import {View} from "@tsed/schema";
-
-export function Vike(path?: string) {
-  return View(`${path || "*"}.vike`);
-}
