+++ packages/third-parties/vike/src/decorators/vite.spec.ts
@@ -3,8 +3,6 @@ import {Store} from "@tsed/core";
 
 import {Vite} from "./vite";
 
-jest.mock("vite");
-
 @Controller("/")
 class ControllerTest {
   @Vite()
