+++ packages/third-parties/vike/src/decorators/vite.ts
