+++ packages/third-parties/vike/src/index.ts
@@ -2,11 +2,11 @@
  * @file Automatically generated by barrelsby.
  */
 
-export * from "./VikeModule";
-export * from "./decorators/vike";
-export * from "./interfaces/VikeConfig";
-export * from "./interfaces/VikeRenderContext";
+export * from "./ViteModule";
+export * from "./decorators/vite";
+export * from "./interfaces/ViteConfig";
+export * from "./interfaces/ViteRenderContext";
 export * from "./interfaces/interfaces";
-export * from "./middlewares/VikeRendererMiddleware";
-export * from "./services/VikeServer";
-export * from "./services/VikeService";
+export * from "./middlewares/ViteRendererMiddleware";
+export * from "./services/ViteServer";
+export * from "./services/ViteService";
