--- packages/third-parties/vike/src/interfaces/VikeConfig.ts
@@ -1,8 +1,0 @@
-import type {Options} from "sirv";
-import type {InlineConfig} from "vite";
-
-export interface VikeConfig extends InlineConfig {
-  enableStream?: boolean;
-  statics?: Omit<Options, "dev">;
-  stateSnapshot?(): unknown;
-}
