--- packages/third-parties/vike/src/interfaces/VikeRenderContext.ts
@@ -1,13 +1,0 @@
-import type {IncomingHttpHeaders} from "http";
-
-export type VikePageProps<Data = Record<string, unknown>> = Data & {view: string};
-
-export interface VikeRenderContext<Data = Record<string, unknown>, Session = Record<string, unknown>> extends Record<string, unknown> {
-  host: string;
-  protocol: string;
-  method: string;
-  url: string;
-  secure: boolean;
-  headers: IncomingHttpHeaders;
-  session?: Session;
-}
