+++ packages/third-parties/vike/src/interfaces/ViteRenderContext.ts
