--- packages/third-parties/vite-ssr-plugin/jest.config.js
@@ -1,14 +1,0 @@
-// For a detailed explanation regarding each configuration property, visit:
-// https://jestjs.io/docs/en/configuration.html
-
-module.exports = {
-  ...require("@tsed/jest-config"),
-  coverageThreshold: {
-    global: {
-      statements: 100,
-      branches: 81.08,
-      functions: 100,
-      lines: 100
-    }
-  }
-};
