+++ packages/third-parties/vite-ssr-plugin/src/index.ts
@@ -1,12 +1,5 @@
-/**
- * @file Automatically generated by barrelsby.
- */
+import {ViteService} from "@tsed/vike";
 
-export * from "./ViteModule";
-export * from "./decorators/vite";
-export * from "./interfaces/ViteConfig";
-export * from "./interfaces/ViteRenderContext";
-export * from "./interfaces/interfaces";
-export * from "./middlewares/ViteRendererMiddleware";
-export * from "./services/ViteServer";
-export * from "./services/ViteService";
+export * from "@tsed/vike";
+
+ViteService.moduleName = "vite-ssr-plugin";
