--- packages/third-parties/vite-ssr-plugin/src/interfaces/interfaces.ts
@@ -1,9 +1,0 @@
-import {ViteConfig} from "./ViteConfig";
-
-declare global {
-  namespace TsED {
-    interface Configuration {
-      vite: ViteConfig;
-    }
-  }
-}
