+++ packages/third-parties/vite-ssr-plugin/tsconfig.json
@@ -28,16 +28,7 @@
   ],
   "references": [
     {
-      "path": "../../platform/common"
-    },
-    {
-      "path": "../../core"
-    },
-    {
-      "path": "../../di"
-    },
-    {
-      "path": "../../specs/exceptions"
+      "path": "../vike"
     }
   ]
 }
